Refactor modules and update documentation
